Move allocForScore to predicate_test.go
